data-preprocessing/requirements.txt
numpy==1.26.4
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
pdfminer.six==20231228
PyMuPDF==1.24.10
pillow==10.4.0
pdfplumber==0.11.4
camelot-py[cv]==0.9.0
spacy==3.7.6
nltk==3.9.1
loguru==0.7.2
pytest==7.2.0